Fix IllegalStateException and freezing screen
Create GameFrame "gameframe" and add "game" to the "gameframe".

When game.start() is called in main, "game" will render graphics.
IllegalStateException is thrown because there is no JFrame for the
"game" canvas to render on.

Add if statement to check if the System's "os.name" return "Linux".

The screen freezes and the game loop reaches approximately 7000 frames
per second during runtime. This problem is only found on Linux operating
system distributions. Get the System's "os.name" property and check if
it returns "Linux." If so, set the System's "sun.java2d.opengl" property
to be true.