Create Ball variables and methods
Initialize public integers "x" and "y" as x-position and y-position of the
ball respectively. Initialize the private Dimension "ballDim" with the
size of the ball.

Create Ball initializer and point "this.x" and "this.y" to the middle of
the frame.

In drawBall(), set the color and draw the rectangle relative to "this.x"
and "this.y" and the ball's dimensions.