Create Paddle variables and methods
Initialize Paddle's final width and height to be used to draw it.
Initialize "x" and "y" as the Paddle's initial drawing position on the
GameFrame's grid.

Create Ball constructor with "player" parameter. Create if statements to
check if "player" is 1 or 2. This is needed for the GameController and
to allow different "x" positions for the paddle is to be drawn
initially.

In drawPaddle(), set paddle's color to be white to be seen on the black
background, and fill the rectangle according to the current paddle's "x"
and "y" position.