Add frame rate and tick counter
Create "lastTime" to check current time at start of run(). This is to be
used for referencing. Create a "timer" variable to keep track of the
current run time in milliseconds. Create "ns" variable for nanosecond
tracking per 60, "frames" for frame per second counter, and "tick" to
count how many times tick() was called. "now" is a reference call for
the current time in nanoseconds once entered in the running loop.

The variable "delta" is used to check how much time has changed over
"ns". The delta loop will check if delta is greater than 1 to which it
will run the tick() method, add 1 to ticks and subtract 1 from "delta".
This is for checking for "delta" changes in between "now" and
"lastTime".

Call a current time in milliseconds and find the difference with the
timer created previously and check if it is greater than 1000 ms.
Display the tick count and frame per second count on the title bar and
reset the frames and ticks to zero.