Implement Swing Timer
Import Timer from swing to use for a swing Timer. Create a Timer "timer"
variable. Create an ActionListener "taskPerformer" that is used as a
parameter for creating the Timer. Initiate the delay value to be 1000
(milliseconds). Call and create the new Timer to have the delay and
taskPerformer. The Timer will be used to update and change the current
frame when a paddle moves or when the ball moves.